Adds EventNormalizer and updates the events
I was missing an important point, events need to be stored !
Since I want them to also act like value objects and not being malformed
I need a way for a third party to handle my events on the fly.

With a normalized format it will be possible to handle all events the
same way. The third party will just gave to use the normalizer to get a
normalized event that it can understand.
When it would like to create an event from some data it will just have
to create a normalized event from it and the normalizer will gave it
back an event.